refactor: split frontend book_detail page into hero/body/merge (#466)
`frontend/src/pages/book_detail.rs` had grown to 827 lines, just over
the ~800 soft cap in `.claude/rules/05-rust-style.md`. Split it
mechanically into a `book_detail/` subdirectory along the natural
section seams:

* `mod.rs` — `BookDetailPage` shell, data-fetch effect, `render_loaded`
  + breadcrumb/series helpers, shared markup primitives, tests.
* `hero.rs` — `BdHeroSection` + `BdTitleCol` + `BdCtaRow`.
* `body.rs` — `BdBodyMain` + `BdRailSection`.
* `merge.rs` — the `#[cfg(not(feature = "mobile"))]` "Merge with…"
  button + dialog + post-merge toast, with mobile stubs returning
  `None`.

No logic or markup changes. All `#[cfg(...)]` guards preserved
bit-exactly; the hero/body components stay platform-agnostic and only
the merge helpers carry the `mobile` gate. `pub use` is not needed —
the existing `pages/mod.rs` re-export of `BookDetailPage` continues to
resolve through the new `mod.rs`, so external call sites see zero
change.

This issue's sister target — `db/src/sync/books.rs` — has already
dropped from 844 → 764 lines via PRs #491, #493, and #494 (unused
column removals), so no further split is needed there.

Closes #466
